// src/types/workoutProgress.types.ts
export type WorkoutProgressMode =
    | "last7"
    | "last30"
    | "currentMonth"
    | "customRange";

export type WorkoutProgressCompareTo =
    | "previous_period"
    | "previous_month"
    | "none";